Created Security utils class to handle login security
